[Add/#15] 폴더링 수정 및 NewWishListVM 추가
